fs: move stream logic to separate file
